+++ src/main/java/suite/rt/planar/Plane.java
@@ -14,10 +14,10 @@ import suite.rt.RayTracer.RtObject;
 public class Plane implements RtObject {
 
 	private Vector normal;
-	private float originIndex;
+	private double originIndex;
 	private Material material;
 
-	public Plane(Vector normal, float originIndex, Material material) {
+	public Plane(Vector normal, double originIndex, Material material) {
 		this.normal = normal;
 		this.originIndex = originIndex;
 		this.material = material;
@@ -25,19 +25,19 @@ public class Plane implements RtObject {
 
 	@Override
 	public List<RayHit> hit(Ray ray) {
-		float denum = Vector.dot(normal, ray.dir);
-		float adv;
+		double denum = Vector.dot(normal, ray.dir);
+		double adv;
 
 		if (MathUtil.epsilon < Math.abs(denum))
 			adv = (originIndex - Vector.dot(normal, ray.startPoint)) / denum;
 		else
-			adv = -1f; // treats as not-hit
+			adv = -1d; // treats as not-hit
 
-		float advance = adv;
+		double advance = adv;
 
 		if (RayTracer.negligibleAdvance < advance) {
 			RayHit rayHit = new RayHit() {
-				public float advance() {
+				public double advance() {
 					return advance;
 				}
